+++ data-access-layer/src/main/java/com/example/sudokudbapi/repositories/PlayedGameRepo.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 import com.example.sudokudbapi.dataModes.PlayedGame;
 
 @Repository
-public interface PlayedGameRepo extends CrudRepository<PlayedGame, Long>{}
+public interface PlayedGameRepo extends CrudRepository<PlayedGame, Integer>{}
