+++ data-access-layer/src/main/java/com/example/sudokudbapi/repositories/UserRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.example.sudokudbapi.dataModes.User;
 
 @Repository
-public interface UserRepo extends CrudRepository<User, Long> {
+public interface UserRepo extends CrudRepository<User, Integer> {
 
     Optional<User> findByUsernameAndPassword(String username, String password);
 }
